pyproject.toml
[tool.poetry]
name = "superset_embedded"
version = "0.1.0"
description = "Backend for embedding superset dashboard into a frontend app"
authors = ["adimyth <aditya@fnp.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
python-dotenv = "^0.21.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
